authentication/requirements.txt
Flask==2.2.3
protobuf==4.22.1
google-auth==2.16.2
google-auth-httplib2==0.1.0 
requests==latest
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1
